requirements.txt
flask
flask-cors
psycopg2-binary
pandas
openpyxl
python-dotenv
gunicorn
numpy==2.2.6
google-generativeai>=0.8.3